test(broker-core): improve exporter tests
