+++ broker-core/src/main/java/io/zeebe/broker/Loggers.java
@@ -29,6 +29,11 @@ public class Loggers {
   public static final Logger WORKFLOW_REPOSITORY_LOGGER =
       new ZbLogger("io.zeebe.broker.workflow.repository");
 
-  public static final Logger EXPORTER_LOGGER = new ZbLogger("io.zeebe.broker.exporter");
   public static final Logger WORKFLOW_PROCESSOR_LOGGER = new ZbLogger("io.zeebe.broker.workflow");
+  public static final Logger EXPORTER_LOGGER = new ZbLogger("io.zeebe.broker.exporter");
+
+  public static final Logger getExporterLogger(String exporterId) {
+    final String loggerName = String.format("io.zeebe.broker.exporter.%s", exporterId);
+    return new ZbLogger(loggerName);
+  }
 }
