+++ broker-core/src/test/java/io/zeebe/broker/exporter/stream/ExporterStreamProcessorTest.java
@@ -17,113 +17,127 @@
  */
 package io.zeebe.broker.exporter.stream;
 
-import static io.zeebe.exporter.api.record.Assertions.assertThat;
 import static io.zeebe.test.util.TestUtil.waitUntil;
-import static io.zeebe.util.buffer.BufferUtil.bufferAsString;
-import static io.zeebe.util.buffer.BufferUtil.wrapString;
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.entry;
-import static org.mockito.Mockito.doAnswer;
-import static org.mockito.Mockito.spy;
+import static org.mockito.Mockito.*;
 
-import io.zeebe.broker.exporter.ExporterObjectMapper;
-import io.zeebe.broker.exporter.record.value.*;
-import io.zeebe.broker.exporter.record.value.deployment.DeployedWorkflowImpl;
-import io.zeebe.broker.exporter.record.value.deployment.DeploymentResourceImpl;
-import io.zeebe.broker.exporter.record.value.job.HeadersImpl;
+import io.zeebe.broker.Loggers;
 import io.zeebe.broker.exporter.repo.ExporterDescriptor;
-import io.zeebe.broker.exporter.stream.ExporterRecord.ExporterPosition;
 import io.zeebe.broker.exporter.util.ControlledTestExporter;
 import io.zeebe.broker.exporter.util.PojoConfigurationExporter;
 import io.zeebe.broker.exporter.util.PojoConfigurationExporter.PojoExporterConfiguration;
 import io.zeebe.broker.logstreams.state.DefaultZeebeDbFactory;
-import io.zeebe.broker.subscription.message.data.MessageSubscriptionRecord;
-import io.zeebe.broker.subscription.message.data.WorkflowInstanceSubscriptionRecord;
 import io.zeebe.broker.util.StreamProcessorControl;
 import io.zeebe.broker.util.StreamProcessorRule;
-import io.zeebe.db.ZeebeDb;
+import io.zeebe.exporter.api.context.Context;
 import io.zeebe.exporter.api.record.Record;
-import io.zeebe.exporter.api.record.RecordValue;
-import io.zeebe.exporter.api.record.value.*;
-import io.zeebe.logstreams.log.LoggedEvent;
-import io.zeebe.msgpack.UnpackedObject;
-import io.zeebe.protocol.BpmnElementType;
-import io.zeebe.protocol.ErrorType;
-import io.zeebe.protocol.VariableDocumentUpdateSemantic;
-import io.zeebe.protocol.impl.record.RecordMetadata;
 import io.zeebe.protocol.impl.record.value.deployment.DeploymentRecord;
-import io.zeebe.protocol.impl.record.value.deployment.ResourceType;
-import io.zeebe.protocol.impl.record.value.incident.IncidentRecord;
-import io.zeebe.protocol.impl.record.value.job.JobBatchRecord;
-import io.zeebe.protocol.impl.record.value.job.JobRecord;
-import io.zeebe.protocol.impl.record.value.message.MessageRecord;
-import io.zeebe.protocol.impl.record.value.variable.VariableDocumentRecord;
-import io.zeebe.protocol.impl.record.value.variable.VariableRecord;
-import io.zeebe.protocol.impl.record.value.workflowinstance.WorkflowInstanceCreationRecord;
-import io.zeebe.protocol.impl.record.value.workflowinstance.WorkflowInstanceRecord;
-import io.zeebe.protocol.intent.*;
-import io.zeebe.test.util.MsgPackUtil;
-import io.zeebe.test.util.collection.Maps;
-import io.zeebe.util.buffer.BufferUtil;
+import io.zeebe.protocol.intent.DeploymentIntent;
 import java.time.Duration;
-import java.time.Instant;
 import java.util.*;
 import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicLong;
-import org.agrona.DirectBuffer;
-import org.agrona.concurrent.UnsafeBuffer;
+import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
+import org.mockito.verification.VerificationWithTimeout;
 
 public class ExporterStreamProcessorTest {
+
   private static final int PARTITION_ID = 1;
-  private static final ExporterObjectMapper OBJECT_MAPPER = new ExporterObjectMapper();
-  private static final Map<String, Object> VARIABLES = Collections.singletonMap("foo", "bar");
-  private static final String VARIABLES_JSON = OBJECT_MAPPER.toJson(VARIABLES);
-  private static final DirectBuffer VARIABLES_MSGPACK =
-      new UnsafeBuffer(OBJECT_MAPPER.toMsgpack(VARIABLES));
-  private static final Map<String, Object> CUSTOM_HEADERS =
-      Collections.singletonMap("workerVersion", 42);
-  private static final DirectBuffer CUSTOM_HEADERS_MSGPACK =
-      new UnsafeBuffer(OBJECT_MAPPER.toMsgpack(CUSTOM_HEADERS));
+
+  private static final String EXPORTER_ID_1 = "exporter-1";
+  private static final String EXPORTER_ID_2 = "exporter-2";
+
+  private static final VerificationWithTimeout TIMEOUT = timeout(5_000);
 
   @Rule
   public StreamProcessorRule rule =
       new StreamProcessorRule(
           PARTITION_ID, DefaultZeebeDbFactory.defaultFactory(ExporterColumnFamilies.class));
 
-  private List<ControlledTestExporter> exporters;
+  private List<ControlledTestExporter> exporters = new ArrayList<>();
+  private List<ExporterDescriptor> exporterDescriptors = new ArrayList<>();
+
   private ExporterStreamProcessorState state;
 
+  @Before
+  public void init() {
+    exporters.clear();
+    exporterDescriptors.clear();
+
+    createExporter(EXPORTER_ID_1, Collections.singletonMap("x", 1));
+    createExporter(EXPORTER_ID_2, Collections.singletonMap("y", 2));
+  }
+
+  private void createExporter(final String exporterId, final Map<String, Object> arguments) {
+    final ControlledTestExporter exporter = spy(new ControlledTestExporter());
+
+    final ExporterDescriptor descriptor =
+        spy(new ExporterDescriptor(exporterId, exporter.getClass(), arguments));
+    doAnswer(c -> exporter).when(descriptor).newInstance();
+
+    exporters.add(exporter);
+    exporterDescriptors.add(descriptor);
+  }
+
+  private StreamProcessorControl startStreamProcessor(
+      List<ExporterDescriptor> exporterDescriptors) {
+    return rule.runStreamProcessor(
+        (db, dbContext) -> {
+          final ExporterStreamProcessor streamProcessor =
+              new ExporterStreamProcessor(
+                  db, db.createContext(), PARTITION_ID, exporterDescriptors);
+
+          state = streamProcessor.getState();
+
+          return streamProcessor;
+        });
+  }
+
   @Test
-  public void shouldConfigureAllExportersProperlyOnStart() throws InterruptedException {
-    // given
-    final Map[] arguments = new Map[] {newConfig("foo", "bar"), newConfig("bar", "foo")};
-    final List<ExporterDescriptor> descriptors = createMockedExporters(arguments);
-
-    final CountDownLatch latch = new CountDownLatch(exporters.size());
-    for (ControlledTestExporter exporter : exporters) {
-      exporter.onOpen(c -> latch.countDown());
-    }
-
+  public void shouldConfigureAllExportersProperlyOnStart() {
     // when
-    final StreamProcessorControl control =
-        rule.initStreamProcessor(
-            (db, dbContext) ->
-                new ExporterStreamProcessor(db, dbContext, PARTITION_ID, descriptors));
-    control.start();
-    assertThat(latch.await(5, TimeUnit.SECONDS)).isTrue();
+    startStreamProcessor(exporterDescriptors);
 
     // then
-    for (int i = 0; i < exporters.size(); i++) {
-      assertThat(exporters.get(i).getContext().getConfiguration().getId())
-          .isEqualTo(descriptors.get(i).getId());
-      assertThat(exporters.get(i).getContext().getConfiguration().getArguments())
-          .isEqualTo(arguments[i]);
-      assertThat(exporters.get(i).getContext().getLogger()).isNotNull();
-      assertThat(exporters.get(i).getController()).isNotNull();
-    }
+    verify(exporters.get(0), TIMEOUT).open(any());
+    verify(exporters.get(1), TIMEOUT).open(any());
+
+    exporters.forEach(
+        exporter -> {
+          assertThat(exporter.getController()).isNotNull();
+          assertThat(exporter.getContext().getLogger()).isNotNull();
+          assertThat(exporter.getContext().getConfiguration()).isNotNull();
+        });
+
+    final Context exporterContext1 = exporters.get(0).getContext();
+    assertThat(exporterContext1.getConfiguration().getId()).isEqualTo(EXPORTER_ID_1);
+    assertThat(exporterContext1.getConfiguration().getArguments())
+        .isEqualTo(Collections.singletonMap("x", 1));
+    assertThat(exporterContext1.getLogger().getName())
+        .isEqualTo(Loggers.getExporterLogger(EXPORTER_ID_1).getName());
+
+    final Context exporterContext2 = exporters.get(1).getContext();
+    assertThat(exporterContext2.getConfiguration().getId()).isEqualTo(EXPORTER_ID_2);
+    assertThat(exporterContext2.getConfiguration().getArguments())
+        .isEqualTo(Collections.singletonMap("y", 2));
+    assertThat(exporterContext2.getLogger().getName())
+        .isEqualTo(Loggers.getExporterLogger(EXPORTER_ID_2).getName());
+  }
+
+  @Test
+  public void shouldCloseAllExportersOnClose() {
+    // given
+    final StreamProcessorControl control = startStreamProcessor(exporterDescriptors);
+
+    // when
+    control.close();
+
+    // then
+    verify(exporters.get(0), TIMEOUT).close();
+    verify(exporters.get(1), TIMEOUT).close();
   }
 
   @Test
@@ -147,10 +161,7 @@ public class ExporterStreamProcessorTest {
         new ExporterDescriptor(
             "instantiateConfiguration", PojoConfigurationExporter.class, arguments);
 
-    rule.runStreamProcessor(
-        (db, dbContext) ->
-            new ExporterStreamProcessor(
-                db, dbContext, PARTITION_ID, Collections.singletonList(descriptor)));
+    startStreamProcessor(Collections.singletonList(descriptor));
 
     // then
     final PojoExporterConfiguration configuration = PojoConfigurationExporter.configuration;
@@ -162,146 +173,11 @@ public class ExporterStreamProcessorTest {
   }
 
   @Test
-  public void shouldCloseAllExportersOnClose() {
-    // given
-    final boolean[] closedExporters = new boolean[] {false, false};
-
-    // when
-    final StreamProcessorControl control =
-        rule.runStreamProcessor(
-            (db, dbContext) -> createStreamProcessor(db, closedExporters.length));
-    exporters.get(0).onClose(() -> closedExporters[0] = true);
-    exporters.get(1).onClose(() -> closedExporters[1] = true);
-
-    // then
-    for (int i = 0; i < exporters.size(); i++) {
-      assertThat(closedExporters[i]).isFalse();
-    }
-
-    // when
-    control.close();
-
-    // then
-    for (int i = 0; i < exporters.size(); i++) {
-      assertThat(closedExporters[i]).isTrue();
-    }
-  }
-
-  @Test
-  public void shouldRestartEachExporterFromCorrectPosition() throws Exception {
-    // given
-    final StreamProcessorControl control =
-        rule.runStreamProcessor((db, dbContext) -> createStreamProcessor(db, 2));
-    final AtomicLong atomicLong = new AtomicLong();
-    control.blockAfterEvent(e -> atomicLong.incrementAndGet() == 2);
-
-    // when
-    final long lowestPosition = writeEvent();
-    final long highestPosition = writeEvent();
-    waitUntil(control::isBlocked);
-    control.unblock();
-
-    // then
-    waitUntil(() -> exporters.get(0).getExportedRecords().size() == 2);
-    waitUntil(() -> exporters.get(1).getExportedRecords().size() == 2);
-
-    // when
-    exporters.get(0).getController().updateLastExportedRecordPosition(highestPosition);
-    exporters.get(1).getController().updateLastExportedRecordPosition(lowestPosition);
-    control.close();
-    control.blockAfterEvent(e -> e.getPosition() == highestPosition);
-    control.start();
-    waitUntil(control::isBlocked);
-
-    // then
-    // should have reprocessed no events
-    assertThat(exporters.get(0).getExportedRecords()).hasSize(0);
-
-    // should have reprocessed the last event twice
-    assertThat(exporters.get(1).getExportedRecords()).hasSize(1);
-    assertThat(exporters.get(1).getExportedRecords().get(0).getPosition())
-        .isEqualTo(highestPosition);
-  }
-
-  @Test
-  public void shouldRecoverPositionsFromLogStream() {
-    // given
-    final List<ExporterDescriptor> descriptors = createMockedExporters(1);
-
-    // when
-    final StreamProcessorControl control =
-        rule.initStreamProcessor(
-            (db, dbContext) -> {
-              final ExporterStreamProcessor processor =
-                  new ExporterStreamProcessor(db, dbContext, PARTITION_ID, descriptors);
-              state = processor.getState();
-              return processor;
-            });
-
-    final long lowestPosition = writeEvent();
-    final long latestPosition = writeExporterEvent(descriptors.get(0).getId(), lowestPosition);
-
-    control.blockAfterEvent(e -> e.getPosition() == latestPosition);
-    control.start();
-    waitUntil(control::isBlocked);
-
-    // then
-    assertThat(state.getPosition(descriptors.get(0).getId())).isEqualTo(lowestPosition);
-  }
-
-  @Test
-  public void shouldRemoveExporterFromState() {
-    // given
-    final List<ExporterDescriptor> descriptors = createMockedExporters(2);
-
-    final StreamProcessorControl control =
-        rule.runStreamProcessor(
-            (db, dbContext) -> {
-              final ExporterStreamProcessor processor =
-                  new ExporterStreamProcessor(db, dbContext, PARTITION_ID, descriptors);
-              state = processor.getState();
-              return processor;
-            });
-
-    final long position = writeEvent();
-
-    waitUntil(() -> exporters.get(0).getExportedRecords().size() == 1);
-    waitUntil(() -> exporters.get(1).getExportedRecords().size() == 1);
-
-    exporters.get(0).getController().updateLastExportedRecordPosition(position);
-    exporters.get(1).getController().updateLastExportedRecordPosition(position);
-
-    waitUntil(() -> state.getPosition(descriptors.get(0).getId()) == position);
-    waitUntil(() -> state.getPosition(descriptors.get(1).getId()) == position);
-
-    control.close();
-
-    // when
-    rule.runStreamProcessor(
-        (db, dbContext) -> {
-          final ExporterStreamProcessor processor =
-              new ExporterStreamProcessor(db, dbContext, PARTITION_ID, descriptors.subList(0, 1));
-          state = processor.getState();
-          return processor;
-        });
-
-    writeEvent();
-    waitUntil(() -> exporters.get(0).getExportedRecords().size() == 2);
-
-    // then
-    assertThat(state.getPosition(descriptors.get(0).getId())).isEqualTo(position);
-    assertThat(state.getPosition(descriptors.get(1).getId()))
-        .isEqualTo(ExporterRecord.POSITION_UNKNOWN);
-  }
-
-  @Test
   public void shouldRetryExportingOnException() {
-    final StreamProcessorControl control =
-        rule.runStreamProcessor((db, dbContext) -> createStreamProcessor(db, 3));
-
+    // given
     final AtomicLong failCount = new AtomicLong(3);
     exporters
-        .get(1)
+        .get(0)
         .onExport(
             e -> {
               if (failCount.getAndDecrement() > 0) {
@@ -309,591 +185,149 @@ public class ExporterStreamProcessorTest {
               }
             });
 
-    // when
-    final long lowestPosition = writeEvent();
-    final long highestPosition = writeEvent();
+    final StreamProcessorControl control = startStreamProcessor(exporterDescriptors);
 
-    control.blockAfterEvent(e -> e.getPosition() == highestPosition);
+    // when
+    final long eventPosition1 = writeEvent();
+    final long eventPosition2 = writeEvent();
+
+    control.blockAfterEvent(e -> e.getPosition() == eventPosition2);
     waitUntil(control::isBlocked);
 
     // then
-    for (final ControlledTestExporter exporter : exporters) {
-      assertThat(exporter.getExportedRecords())
-          .extracting("position")
-          .containsExactly(lowestPosition, highestPosition);
-    }
+    assertThat(exporters.get(0).getExportedRecords())
+        .extracting(Record::getPosition)
+        .containsExactly(eventPosition1, eventPosition2);
+    assertThat(exporters.get(1).getExportedRecords())
+        .extracting(Record::getPosition)
+        .containsExactly(eventPosition1, eventPosition2);
   }
 
   @Test
   public void shouldExecuteScheduledTask() throws Exception {
     // given
-    final List<ExporterDescriptor> mockedExporters = createMockedExporters(1);
     final CountDownLatch timerTriggerLatch = new CountDownLatch(1);
     final CountDownLatch timerScheduledLatch = new CountDownLatch(1);
     final Duration delay = Duration.ofSeconds(10);
 
-    final ControlledTestExporter controlledTestExporter = exporters.get(0);
-    controlledTestExporter.onExport(
-        r -> {
-          controlledTestExporter.getController().scheduleTask(delay, timerTriggerLatch::countDown);
-          timerScheduledLatch.countDown();
-        });
-
-    rule.runStreamProcessor((db, dbContext) -> createStreamProcessor(db, mockedExporters));
-    writeEvent();
-    timerScheduledLatch.await();
+    exporters
+        .get(0)
+        .onExport(
+            r -> {
+              exporters.get(0).getController().scheduleTask(delay, timerTriggerLatch::countDown);
+              timerScheduledLatch.countDown();
+            });
 
     // when
-    rule.getClock().addTime(delay.plusSeconds(20));
-    timerTriggerLatch.await();
+    startStreamProcessor(exporterDescriptors);
+
+    writeEvent();
+
+    timerScheduledLatch.await(5, TimeUnit.SECONDS);
+
+    rule.getClock().addTime(delay);
 
     // then
+    assertThat(timerTriggerLatch.await(5, TimeUnit.SECONDS)).isTrue();
   }
 
   @Test
-  public void shouldNotExportExporterRecords() {
+  public void shouldRecoverPositionsFromState() {
     // given
-    final List<ExporterDescriptor> descriptors = createMockedExporters(1);
+    final StreamProcessorControl control = startStreamProcessor(exporterDescriptors);
 
-    // when
-    final StreamProcessorControl control =
-        rule.initStreamProcessor(
-            (db, dbContext) ->
-                new ExporterStreamProcessor(db, dbContext, PARTITION_ID, descriptors));
-    final long lowestPosition = writeEvent();
-    final long latestPosition = writeExporterEvent(descriptors.get(0).getId(), lowestPosition);
+    final long eventPosition1 = writeEvent();
+    final long eventPosition2 = writeEvent();
 
-    control.blockAfterEvent(e -> e.getPosition() == latestPosition);
+    waitUntil(() -> exporters.get(0).getExportedRecords().size() == 2);
+    waitUntil(() -> exporters.get(1).getExportedRecords().size() == 2);
+
+    exporters.get(0).getController().updateLastExportedRecordPosition(eventPosition2);
+    exporters.get(1).getController().updateLastExportedRecordPosition(eventPosition1);
+
+    control.close();
+
+    exporters.get(0).getExportedRecords().clear();
+    exporters.get(1).getExportedRecords().clear();
+
+    // then
+    control.blockAfterEvent(e -> e.getPosition() >= eventPosition2);
     control.start();
     waitUntil(control::isBlocked);
 
     // then
-    assertThat(exporters.get(0).getExportedRecords()).hasSize(1);
-    assertThat(exporters.get(0).getExportedRecords().get(0).getPosition())
-        .isEqualTo(lowestPosition);
-  }
-
-  @Test
-  public void shouldExportDeploymentEvent() {
-    // given
-    final String resourceName = "resource";
-    final ResourceType resourceType = ResourceType.BPMN_XML;
-    final DirectBuffer resource = wrapString("contents");
-    final String bpmnProcessId = "testProcess";
-    final long workflowKey = 123;
-    final int workflowVersion = 12;
-
-    final DeploymentRecord record = new DeploymentRecord();
-    record
-        .resources()
-        .add()
-        .setResourceName(wrapString(resourceName))
-        .setResourceType(resourceType)
-        .setResource(resource);
-    record
-        .workflows()
-        .add()
-        .setBpmnProcessId(wrapString(bpmnProcessId))
-        .setKey(workflowKey)
-        .setResourceName(wrapString(resourceName))
-        .setVersion(workflowVersion);
-
-    final DeploymentRecordValue expectedRecordValue =
-        new io.zeebe.broker.exporter.record.value.DeploymentRecordValueImpl(
-            OBJECT_MAPPER,
-            Collections.singletonList(
-                new DeployedWorkflowImpl(
-                    bpmnProcessId, resourceName, workflowKey, workflowVersion)),
-            Collections.singletonList(
-                new DeploymentResourceImpl(
-                    BufferUtil.bufferAsArray(resource),
-                    io.zeebe.exporter.api.record.value.deployment.ResourceType.BPMN_XML,
-                    resourceName)));
-
-    // then
-    assertRecordExported(DeploymentIntent.CREATE, record, expectedRecordValue);
-  }
-
-  @Test
-  public void shouldExportIncidentRecord() {
-    // given
-    final long elementInstanceKey = 34;
-    final long workflowInstanceKey = 10;
-    final String elementId = "activity";
-    final String bpmnProcessId = "process";
-    final String errorMessage = "error";
-    final ErrorType errorType = ErrorType.IO_MAPPING_ERROR;
-    final long jobKey = 123;
-
-    final IncidentRecord record =
-        new IncidentRecord()
-            .setElementInstanceKey(elementInstanceKey)
-            .setWorkflowInstanceKey(workflowInstanceKey)
-            .setElementId(wrapString(elementId))
-            .setBpmnProcessId(wrapString(bpmnProcessId))
-            .setErrorMessage(errorMessage)
-            .setErrorType(errorType)
-            .setJobKey(jobKey)
-            .setVariableScopeKey(elementInstanceKey);
-
-    final IncidentRecordValue recordValue =
-        new IncidentRecordValueImpl(
-            OBJECT_MAPPER,
-            errorType.name(),
-            errorMessage,
-            bpmnProcessId,
-            elementId,
-            workflowInstanceKey,
-            elementInstanceKey,
-            jobKey,
-            elementInstanceKey);
-
-    // then
-    assertRecordExported(IncidentIntent.CREATED, record, recordValue);
-  }
-
-  @Test
-  public void shouldExportJobRecord() {
-    // given
-    final String worker = "myWorker";
-    final String type = "myType";
-    final int retries = 12;
-    final int deadline = 13;
-
-    final String bpmnProcessId = "test-process";
-    final int workflowKey = 13;
-    final int workflowDefinitionVersion = 12;
-    final int workflowInstanceKey = 1234;
-    final String elementId = "activity";
-    final int activityInstanceKey = 123;
-
-    final JobRecord record =
-        new JobRecord()
-            .setWorker(wrapString(worker))
-            .setType(wrapString(type))
-            .setVariables(VARIABLES_MSGPACK)
-            .setRetries(retries)
-            .setDeadline(deadline)
-            .setErrorMessage("failed message");
-    record
-        .getHeaders()
-        .setBpmnProcessId(wrapString(bpmnProcessId))
-        .setWorkflowKey(workflowKey)
-        .setWorkflowDefinitionVersion(workflowDefinitionVersion)
-        .setWorkflowInstanceKey(workflowInstanceKey)
-        .setElementId(wrapString(elementId))
-        .setElementInstanceKey(activityInstanceKey);
-
-    record.setCustomHeaders(CUSTOM_HEADERS_MSGPACK);
-
-    final JobRecordValue recordValue =
-        new io.zeebe.broker.exporter.record.value.JobRecordValueImpl(
-            OBJECT_MAPPER,
-            VARIABLES_JSON,
-            type,
-            worker,
-            Instant.ofEpochMilli(deadline),
-            new HeadersImpl(
-                bpmnProcessId,
-                elementId,
-                activityInstanceKey,
-                workflowInstanceKey,
-                workflowKey,
-                workflowDefinitionVersion),
-            CUSTOM_HEADERS,
-            retries,
-            "failed message");
-
-    // then
-    assertRecordExported(JobIntent.CREATED, record, recordValue);
-  }
-
-  @Test
-  public void shouldExportMessageRecord() {
-    // given
-    final String correlationKey = "test-key";
-    final String messageName = "test-message";
-    final long timeToLive = 12;
-    final String messageId = "test-id";
-
-    final MessageRecord record =
-        new MessageRecord()
-            .setCorrelationKey(wrapString(correlationKey))
-            .setName(wrapString(messageName))
-            .setVariables(VARIABLES_MSGPACK)
-            .setTimeToLive(timeToLive)
-            .setMessageId(wrapString(messageId));
-
-    final MessageRecordValue recordValue =
-        new io.zeebe.broker.exporter.record.value.MessageRecordValueImpl(
-            OBJECT_MAPPER, VARIABLES_JSON, messageName, messageId, correlationKey, timeToLive);
-
-    // then
-    assertRecordExported(MessageIntent.PUBLISHED, record, recordValue);
-  }
-
-  @Test
-  public void shouldExportMessageSubscriptionRecord() {
-    // given
-    final long activityInstanceKey = 1L;
-    final String messageName = "name";
-    final long workflowInstanceKey = 1L;
-    final String correlationKey = "key";
-
-    final MessageSubscriptionRecord record =
-        new MessageSubscriptionRecord()
-            .setElementInstanceKey(activityInstanceKey)
-            .setMessageName(wrapString(messageName))
-            .setWorkflowInstanceKey(workflowInstanceKey)
-            .setCorrelationKey(wrapString(correlationKey));
-
-    final MessageSubscriptionRecordValue recordValue =
-        new MessageSubscriptionRecordValueImpl(
-            OBJECT_MAPPER, messageName, correlationKey, workflowInstanceKey, activityInstanceKey);
-
-    // then
-    assertRecordExported(MessageSubscriptionIntent.CORRELATE, record, recordValue);
-  }
-
-  @Test
-  public void shouldExportWorkflowInstanceRecord() {
-    // given
-    final String bpmnProcessId = "test-process";
-    final int workflowKey = 13;
-    final int version = 12;
-    final int workflowInstanceKey = 1234;
-    final String elementId = "activity";
-    final int flowScopeKey = 123;
-    final BpmnElementType bpmnElementType = BpmnElementType.SERVICE_TASK;
-
-    final WorkflowInstanceRecord record =
-        new WorkflowInstanceRecord()
-            .setElementId(elementId)
-            .setBpmnElementType(bpmnElementType)
-            .setBpmnProcessId(wrapString(bpmnProcessId))
-            .setVersion(version)
-            .setWorkflowKey(workflowKey)
-            .setWorkflowInstanceKey(workflowInstanceKey)
-            .setFlowScopeKey(flowScopeKey);
-
-    final WorkflowInstanceRecordValue recordValue =
-        new WorkflowInstanceRecordValueImpl(
-            OBJECT_MAPPER,
-            bpmnProcessId,
-            elementId,
-            version,
-            workflowKey,
-            workflowInstanceKey,
-            flowScopeKey,
-            bpmnElementType);
-
-    // then
-    assertRecordExported(WorkflowInstanceIntent.ELEMENT_ACTIVATING, record, recordValue);
-  }
-
-  @Test
-  public void shouldExportWorkflowInstanceSubscriptionRecord() {
-    // given
-    final long activityInstanceKey = 123;
-    final String messageName = "test-message";
-    final int subscriptionPartitionId = 2;
-    final long workflowInstanceKey = 1345;
-
-    final WorkflowInstanceSubscriptionRecord record =
-        new WorkflowInstanceSubscriptionRecord()
-            .setElementInstanceKey(activityInstanceKey)
-            .setMessageName(wrapString(messageName))
-            .setSubscriptionPartitionId(subscriptionPartitionId)
-            .setWorkflowInstanceKey(workflowInstanceKey)
-            .setVariables(VARIABLES_MSGPACK);
-
-    final WorkflowInstanceSubscriptionRecordValue recordValue =
-        new WorkflowInstanceSubscriptionRecordValueImpl(
-            OBJECT_MAPPER, VARIABLES_JSON, messageName, workflowInstanceKey, activityInstanceKey);
-
-    // then
-    assertRecordExported(WorkflowInstanceSubscriptionIntent.OPENED, record, recordValue);
-  }
-
-  @Test
-  public void shouldExportJobBatchRecord() {
-    // given
-    final int amount = 1;
-    final long timeout = 2L;
-    final String type = "type";
-    final String worker = "worker";
-
-    final JobBatchRecord record =
-        new JobBatchRecord()
-            .setMaxJobsToActivate(amount)
-            .setTimeout(timeout)
-            .setType(type)
-            .setWorker(worker)
-            .setTruncated(true);
-
-    record.jobKeys().add().setValue(3L);
-    final JobRecord jobRecord = record.jobs().add();
-
-    final String bpmnProcessId = "test-process";
-    final int workflowKey = 13;
-    final int workflowDefinitionVersion = 12;
-    final int workflowInstanceKey = 1234;
-    final String activityId = "activity";
-    final int activityInstanceKey = 123;
-
-    jobRecord
-        .setWorker(wrapString(worker))
-        .setType(wrapString(type))
-        .setVariables(VARIABLES_MSGPACK)
-        .setRetries(3)
-        .setErrorMessage("failed message")
-        .setDeadline(1000L);
-
-    jobRecord
-        .getHeaders()
-        .setBpmnProcessId(wrapString(bpmnProcessId))
-        .setWorkflowKey(workflowKey)
-        .setWorkflowDefinitionVersion(workflowDefinitionVersion)
-        .setWorkflowInstanceKey(workflowInstanceKey)
-        .setElementId(wrapString(activityId))
-        .setElementInstanceKey(activityInstanceKey);
-
-    final JobRecordValueImpl jobRecordValue =
-        new JobRecordValueImpl(
-            OBJECT_MAPPER,
-            VARIABLES_JSON,
-            type,
-            worker,
-            Instant.ofEpochMilli(1000L),
-            new HeadersImpl(
-                bpmnProcessId,
-                activityId,
-                activityInstanceKey,
-                workflowInstanceKey,
-                workflowKey,
-                workflowDefinitionVersion),
-            Collections.EMPTY_MAP,
-            3,
-            "failed message");
-
-    final JobBatchRecordValueImpl recordValue =
-        new JobBatchRecordValueImpl(
-            OBJECT_MAPPER,
-            type,
-            worker,
-            Duration.ofMillis(timeout),
-            amount,
-            Arrays.asList(3L),
-            Arrays.asList(jobRecordValue),
-            record.getTruncated());
-
-    // then
-    assertRecordExported(JobBatchIntent.ACTIVATED, record, recordValue);
-  }
-
-  @Test
-  public void shouldExportVariableRecord() {
-    // given
-    final String name = "x";
-    final String value = "1";
-    final long scopeKey = 3;
-    final long workflowInstanceKey = 2;
-    final long workflowKey = 4;
-
-    final VariableRecord record =
-        new VariableRecord()
-            .setName(wrapString(name))
-            .setValue(MsgPackUtil.asMsgPack(value))
-            .setScopeKey(scopeKey)
-            .setWorkflowInstanceKey(workflowInstanceKey)
-            .setWorkflowKey(workflowKey);
-
-    final VariableRecordValue recordValue =
-        new VariableRecordValueImpl(
-            OBJECT_MAPPER, name, value, scopeKey, workflowInstanceKey, workflowKey);
-
-    // then
-    assertRecordExported(VariableIntent.CREATED, record, recordValue);
-  }
-
-  @Test
-  public void shouldExportVariableDocumentRecord() {
-    // given
-    final long scopeKey = 1L;
-    final VariableDocumentUpdateSemantic updateSemantics = VariableDocumentUpdateSemantic.LOCAL;
-    final Map<String, Object> document = Maps.of(entry("foo", "bar"), entry("baz", "boz"));
-
-    final VariableDocumentRecord record =
-        new VariableDocumentRecord()
-            .setScopeKey(scopeKey)
-            .setUpdateSemantics(updateSemantics)
-            .setDocument(MsgPackUtil.asMsgPack(document));
-
-    final VariableDocumentRecordValue recordValue =
-        new VariableDocumentRecordValueImpl(OBJECT_MAPPER, scopeKey, updateSemantics, document);
-
-    // then
-    assertRecordExported(VariableDocumentIntent.UPDATED, record, recordValue);
-  }
-
-  @Test
-  public void shouldExportWorkflowInstanceCreationRecord() {
-    // given
-    final String processId = "process";
-    final long key = 1L;
-    final int version = 1;
-    final long instanceKey = 2L;
-    final Map<String, Object> variables = Maps.of(entry("foo", "bar"), entry("baz", "boz"));
-
-    final WorkflowInstanceCreationRecord record =
-        new WorkflowInstanceCreationRecord()
-            .setBpmnProcessId(processId)
-            .setKey(key)
-            .setVersion(version)
-            .setVariables(MsgPackUtil.asMsgPack(variables))
-            .setInstanceKey(instanceKey);
-
-    final WorkflowInstanceCreationRecordValue recordValue =
-        new WorkflowInstanceCreationRecordValueImpl(
-            OBJECT_MAPPER, processId, version, key, instanceKey, variables);
-
-    // then
-    assertRecordExported(WorkflowInstanceCreationIntent.CREATED, record, recordValue);
+    assertThat(exporters.get(0).getExportedRecords()).hasSize(0);
+    assertThat(exporters.get(1).getExportedRecords())
+        .extracting(Record::getPosition)
+        .hasSize(1)
+        .contains(eventPosition2);
   }
 
   @Test
   public void shouldUpdateLastExportedPositionOnClose() {
     // given
-    final StreamProcessorControl control =
-        rule.initStreamProcessor((db, dbContext) -> createStreamProcessor(db, 1));
-    control.start();
+    final StreamProcessorControl control = startStreamProcessor(exporterDescriptors);
 
-    final long firstPosition = writeEvent();
-    final ControlledTestExporter firstExporter = exporters.get(0);
-    waitUntil(() -> firstExporter.getExportedRecords().size() == 1);
-    firstExporter.onClose(
-        () -> firstExporter.getController().updateLastExportedRecordPosition(firstPosition));
-    control.close();
+    final long eventPosition1 = writeEvent();
+    final long eventPosition2 = writeEvent();
+
+    waitUntil(() -> exporters.get(0).getExportedRecords().size() == 2);
+    waitUntil(() -> exporters.get(1).getExportedRecords().size() == 2);
+
+    exporters
+        .get(0)
+        .onClose(
+            () ->
+                exporters.get(0).getController().updateLastExportedRecordPosition(eventPosition1));
 
     // when
-    final long secondPosition = writeEvent();
-    control.blockAfterEvent(e -> e.getPosition() == secondPosition);
+    control.close();
+
+    exporters.get(0).getExportedRecords().clear();
+    exporters.get(1).getExportedRecords().clear();
+
+    control.blockAfterEvent(e -> e.getPosition() == eventPosition2);
     control.start();
     waitUntil(control::isBlocked);
 
     // then
-    final List<Record> records = firstExporter.getExportedRecords();
-    assertThat(records).hasSize(1);
-    assertThat(records.get(0).getPosition()).isEqualTo(firstPosition);
-
-    final ControlledTestExporter secondExporter = exporters.get(0);
-    assertThat(firstExporter).isNotEqualTo(secondExporter);
-
-    final List<Record> secondRecords = secondExporter.getExportedRecords();
-    assertThat(secondRecords).hasSize(1);
-    assertThat(secondRecords.get(0).getPosition()).isEqualTo(secondPosition);
+    assertThat(exporters.get(0).getExportedRecords())
+        .extracting(Record::getPosition)
+        .hasSize(1)
+        .contains(eventPosition2);
+    assertThat(exporters.get(1).getExportedRecords())
+        .extracting(Record::getPosition)
+        .hasSize(2)
+        .contains(eventPosition1, eventPosition2);
   }
 
-  private ExporterStreamProcessor createStreamProcessor(
-      ZeebeDb db, final List<ExporterDescriptor> exporterDescriptors) {
-    return new ExporterStreamProcessor(db, db.createContext(), PARTITION_ID, exporterDescriptors);
-  }
+  @Test
+  public void shouldRemoveExporterFromState() {
+    // given
+    final StreamProcessorControl control = startStreamProcessor(exporterDescriptors);
 
-  private ExporterStreamProcessor createStreamProcessor(ZeebeDb db, final int count) {
-    return new ExporterStreamProcessor(
-        db, db.createContext(), PARTITION_ID, createMockedExporters(count));
-  }
+    final long eventPosition = writeEvent();
 
-  private List<ExporterDescriptor> createMockedExporters(final int count) {
-    return createMockedExporters(count, new Map[0]);
-  }
+    waitUntil(() -> exporters.get(0).getExportedRecords().size() == 1);
+    waitUntil(() -> exporters.get(1).getExportedRecords().size() == 1);
 
-  private List<ExporterDescriptor> createMockedExporters(final Map... arguments) {
-    return createMockedExporters(arguments.length, arguments);
-  }
+    exporters.get(0).getController().updateLastExportedRecordPosition(eventPosition);
+    exporters.get(1).getController().updateLastExportedRecordPosition(eventPosition);
 
-  @SuppressWarnings("unchecked")
-  private List<ExporterDescriptor> createMockedExporters(final int count, final Map[] arguments) {
-    final List<ExporterDescriptor> descriptors = new ArrayList<>(count);
-    exporters = new ArrayList<>(count);
+    control.close();
 
-    for (int i = 0; i < count; i++) {
-      final Map args = arguments.length > 0 ? arguments[i] : null;
-      final ControlledTestExporter exporter = spy(new ControlledTestExporter());
-      final ExporterDescriptor descriptor =
-          spy(new ExporterDescriptor(String.valueOf(i), exporter.getClass(), args));
-      doAnswer(c -> exporter).when(descriptor).newInstance();
+    // when
+    startStreamProcessor(Collections.singletonList(exporterDescriptors.get(0)));
 
-      exporters.add(exporter);
-      descriptors.add(descriptor);
-    }
+    verify(exporters.get(0), TIMEOUT.times(2)).open(any());
 
-    return descriptors;
+    // then
+    assertThat(state.getPosition(EXPORTER_ID_1)).isEqualTo(eventPosition);
+    assertThat(state.getPosition(EXPORTER_ID_2)).isEqualTo(ExporterRecord.POSITION_UNKNOWN);
   }
 
   private long writeEvent() {
     final DeploymentRecord event = new DeploymentRecord();
     return rule.writeEvent(DeploymentIntent.CREATED, event);
-  }
-
-  private long writeExporterEvent(final String id, final long position) {
-    final ExporterRecord event = new ExporterRecord();
-    final ExporterPosition exporterPosition = event.getPositions().add();
-    exporterPosition.setId(id);
-    exporterPosition.setPosition(position);
-
-    return rule.writeEvent(ExporterIntent.EXPORTED, event);
-  }
-
-  private Map<String, Object> newConfig(final String... pairs) {
-    final Map<String, Object> config = new HashMap<>();
-
-    for (int i = 0; i < pairs.length; i += 2) {
-      config.put(pairs[i], pairs[i + 1]);
-    }
-
-    return config;
-  }
-
-  private void assertRecordExported(
-      final Intent intent, final UnpackedObject record, final RecordValue expectedRecordValue) {
-    // setup stream processor
-    final StreamProcessorControl control =
-        rule.initStreamProcessor((db, dbContext) -> createStreamProcessor(db, 1));
-
-    // write event
-    final long position = rule.writeEvent(intent, record);
-
-    // wait for event
-    control.blockAfterEvent(e -> e.getPosition() == position);
-    control.start();
-    waitUntil(control::isBlocked);
-
-    // assert exported record
-    final List<Record> exportedRecords = exporters.get(0).getExportedRecords();
-    assertThat(exportedRecords).hasSize(1);
-
-    final Record actualRecord = exportedRecords.get(0);
-    final LoggedEvent loggedEvent = rule.events().withPosition(position);
-    final RecordMetadata metadata = new RecordMetadata();
-
-    assertThat(actualRecord)
-        .hasPosition(loggedEvent.getPosition())
-        .hasSourceRecordPosition(loggedEvent.getSourceEventPosition())
-        .hasProducerId(loggedEvent.getProducerId())
-        .hasKey(loggedEvent.getKey())
-        .hasTimestamp(Instant.ofEpochMilli(loggedEvent.getTimestamp()));
-
-    loggedEvent.readMetadata(metadata);
-
-    assertThat(actualRecord.getMetadata())
-        .hasIntent(metadata.getIntent())
-        .hasPartitionId(PARTITION_ID)
-        .hasRecordType(metadata.getRecordType())
-        .hasRejectionType(metadata.getRejectionType())
-        .hasRejectionReason(bufferAsString(metadata.getRejectionReason()))
-        .hasValueType(metadata.getValueType());
-
-    assertThat(actualRecord).hasValue(expectedRecordValue);
   }
 }
